requirements.txt

pandas
httpx
supabase
google-genai
groq
anthropic
openai
reflex-google-auth
reflex-clerk-api @ git+https://github.com/reflex-dev/reflex-clerk-api.git
postgrest
agno
typer==0.12.3
stripe
fastapi
apscheduler
PyGithub
cryptography
twilio
pyjwt
together
lightgbm
xgboost
nltk
rake-nltk
joblib
numpy
pymetasploit3
docker
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
reflex
sentence-transformers
spacy
scikit-learn
linear-py